Incidence and incidence difference statistics from trinomial counts of HIV and recency

Incidence and incidence difference statistics from trinomial counts of HIV and recency

inccounts(
  N,
  N_H,
  N_testR,
  N_R,
  DE_H = 1,
  DE_R = 1,
  BS_Count = 10000,
  Boot = FALSE,
  alpha = 0.05,
  BMest = "same.test",
  MDRI,
  RSE_MDRI,
  FRR,
  RSE_FRR,
  BigT = 730,
  Covar_HR = 0,
  debug = FALSE
)

Arguments

N
Counts of total survey sample size(s) (vector/integer).

N_H
Number of HIV positive found in survey(s) (vector/integer).

N_testR
Number tested for recency in survey(s) (vector/integer).

N_R
Number of recent cases in survey(s) (vector/integer).

DE_H
Design effect of HIV prevalence test (vector/numeric), greater than or equal to 1. If multiple surveys are entered but only one design effect is specified, function assumes entered design effect is identical for both surveys.

DE_R
Design effect of recency test (vector/numeric), greater than or equal to 1. If multiple surveys are entered but only one design effect is specified, function assumes entered design effect is identical for both surveys.

BS_Count
Specifies number of bootstrap samples for bootstrapped confidence intervals of incidence.

Boot
True/False variable indicating whether variance of point estimates is to be calculated by Empirical Bootstrapping (TRUE) or Delta Method (FALSE), the default setting.

alpha
test rejection threshold.

BMest
Biomarker estimation by one the 3 options 'same.test'(=default), 'FRR.indep', 'MDRI.FRR.indep' (string).

MDRI
mean duration of recent infection [days] (vector/integer).

RSE_MDRI
Relative standard error of MDRI [days] (vector/integer).

FRR
False recent rate (vector/integer).

RSE_FRR
Relative standard error of FRR (vector/integer).

BigT
Cut point in days of recency used in biomarker assay to test recency in a given prevalence survey.

Covar_HR
Covariance of probability of being postive and being categorized recent from survey (or as a vector for multiple surveys).

debug
Enable debugging mode (browser)

Value

Incidence estimate, annual risk of infection, confidence interval, relative standard error of estimate and of assay characteristics MDRI and FRR. If multiple surveys are entered, function returns said results, as well as estimates of incidence differences, confidence intervals of differences, difference relative standard errors, and p-values testing the hypothesis that the difference in incidence measures are zero. Theoretical relative standard error of incidence and incidence difference at infinite sample size is returned only if Boot = FALSE, as that calculation relies on the asymptotic behavior of components of the Delta method approximation, and is not calculable from bootstrapped values.

Details

Implements assay-based incidence estimation through cross-sectional prevalence and recency of infection tests as described by Kassanjee, et al. 'A new general biomarker-based incidence estimator,' Epidemiology (2012). Function parameters must be specified to include assay test characteristics and survey results as proportions. Confidence intervals are computed via Delta method approximation, except when Boot=TRUE is specified, in which case confidence intervals are generated by empirical bootstrap resampling. Inputs must be in appropriate ranges for appropriate units. Extreme input values may make calculation impossible, and if entered will elicit error notices. The package contains long form documentation in the form of vignettes that cover the use of the main fucntions. Use browseVignettes(package="inctools") to access them.

Examples

inccounts(N = c(5000) ,N_H = 1000, N_testR = 1000, N_R = 70,
Boot = FALSE, BMest = 'MDRI.FRR.indep', MDRI = 200, RSE_MDRI = 0.05,
FRR = 0.01, RSE_FRR = 0.2, BigT = 730)

#> Warning: 'same.test' is the only BMest option for a single test

#> $Incidence.Statistics
#> # A tibble: 1 x 5
#>   Incidence     CI_LB     CI_UB      RSE RSE_InfSS
#>       <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1 0.0284315 0.0200376 0.0368254 0.150631 0.0579346
#> 
#> $Annual.Risk.of.Infection
#> # A tibble: 1 x 3
#>         ARI ARI.CI_LB ARI.CI_UB
#>       <dbl>     <dbl>     <dbl>
#> 1 0.0280311 0.0198382 0.0361556
#> 
#> $MDRI.CI
#>      CI_LB    CI_UB
#> 1 180.4004 219.5996
#> 
#> $FRR.CI
#> # A tibble: 1 x 2
#>        CI_LB     CI_UB
#>        <dbl>     <dbl>
#> 1 0.00608007 0.0139199
#> 

inccounts(N = c(4000,4000,4050) ,N_H = c(1010,1000,900),
N_testR = c(1000,1000,880), N_R = c(60,70,50), Boot = TRUE,
BMest = 'same.test', MDRI = 210, RSE_MDRI = 0.05, FRR = 0.005,
RSE_FRR = 0.19, BigT = 700)

#> $Incidence.Statistics
#> # A tibble: 3 x 7
#>   survey Incidence     CI_LB     CI_UB      RSE   Cov.PrevH.I Cor.PrevH.I
#>    <int>     <dbl>     <dbl>     <dbl>    <dbl>         <dbl>       <dbl>
#> 1      1 0.0328612 0.0237975 0.0429081 0.149904 0.00000788622    0.233767
#> 2      2 0.0383232 0.0285442 0.0494633 0.139472 0.00000924528    0.254177
#> 3      3 0.0261869 0.0179646 0.0350674 0.166763 0.00000733899    0.255692
#> 
#> $Incidence.Difference.Statistics
#> # A tibble: 6 x 6
#>   compare        Diff       CI_LB       CI_UB       RSE     p_value
#>   <chr>         <dbl>       <dbl>       <dbl>     <dbl>       <dbl>
#> 1 1 vs 2  -0.00546207 -0.0197030   0.00855689 1.32921   4.51853e- 1
#> 2 1 vs 3   0.00667426 -0.00622280  0.0196958  0.993099  3.13959e- 1
#> 3 2 vs 1   0.00546207 -0.00855689  0.0197030  1.32921   4.51853e- 1
#> 4 2 vs 3   0.0121363  -0.00132157  0.0259419  0.572606  8.07411e- 2
#> 5 3 vs 1  -0.00667426 -0.0196958   0.00622280 0.993099  3.13959e- 1
#> 6 3 vs 2  -0.0121363  -0.0143703  -0.0105237  0.0808346 3.75422e-35
#> 
#> $MDRI.CI
#>      CI_LB    CI_UB
#> 1 189.4204 230.5796
#> 
#> $FRR.CI
#> # A tibble: 1 x 2
#>        CI_LB      CI_UB
#>        <dbl>      <dbl>
#> 1 0.00313803 0.00686197
#> 

inccounts(N = c(4000,4000) ,N_H = c(1050,1090),
N_testR = c(1000,1000), N_R = c(60,67), Boot = FALSE, BMest = 'FRR.indep',
MDRI = 220, RSE_MDRI = 0.05, FRR = c(0.005,0.005), RSE_FRR = 0.19,
BigT = 610)

#> Warning: Less than 95% of HIV-positive subjects have recency test results available

#> $Incidence.Statistics
#> # A tibble: 2 x 6
#>   survey Incidence     CI_LB     CI_UB      RSE RSE_InfSS
#>    <int>     <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#> 1      1 0.0329580 0.0232213 0.0426947 0.150732 0.0527636
#> 2      2 0.0390981 0.0281923 0.0500039 0.142316 0.0522575
#> 
#> $Incidence.Difference.Statistics
#> # A tibble: 2 x 8
#>   compare        Diff       CI_LB      CI_UB     RSE RSE_InfSS  p_value
#>   <chr>         <dbl>       <dbl>      <dbl>   <dbl>     <dbl>    <dbl>
#> 1 1 vs 2  -0.00614013 -0.0198620  0.00758172 1.14022  0.123302 0.380472
#> 2 2 vs 1   0.00614013 -0.00758172 0.0198620  1.14022  0.123302 0.380472
#> # … with 1 more variable: p_value_InfSS <dbl>
#> 
#> $MDRI.CI
#>      CI_LB    CI_UB
#> 1 198.4404 241.5596
#> 
#> $FRR.CI
#> # A tibble: 2 x 2
#>        CI_LB      CI_UB
#>        <dbl>      <dbl>
#> 1 0.00313803 0.00686197
#> 2 0.00313803 0.00686197
#>