Precision and Required Sample Size

Calculate precision of incidence estimate and required sample size for a given precision of the incidence estimate.

incprecision(
  I,
  RSE_I,
  PrevH,
  CR,
  MDRI,
  RSE_MDRI,
  FRR,
  RSE_FRR,
  BigT = 730,
  DE_H = 1,
  DE_R = 1,
  n = "out",
  step = 5
)

Arguments

I
Expected Incidence.

RSE_I
Relative Standard Error of Incidence Estimate. If this is the desired output, set to "out".

PrevH
Prevalence of HIV.

CR
Coverage rate: probability (0-1) of being tested for recency when positive for HIV.

MDRI
mean duration of recent infection in days (vector/integer).

RSE_MDRI
Relative standard error of MDRI (vector/integer).

FRR
False recent rate (vector/integer).

RSE_FRR
Relative standard error of FRR (vector/integer).

BigT
post-infection time cut-off for true vs. false recency. Default is 730 days.

DE_H
Design effect of HIV prevalence test (vector/integer).

DE_R
Design effect of recency test (vector/integer).

n
Sample Size: Set to a hypothetical value if the desired output is RSE_I, othewise set to "out" to obtain required sample size.

step
number of steps between minimum I and maximum I in the calculation of a range of output.

Value

Either sample size necessary for a given precision under a given set of testing characteristics and a hypothetical prevalence/incidence scenario, or precision under a particular sample size scenario, with a given hypothetical prevalence/incidence scenario.

Details

The package contains long form documentation in the form of vignettes that cover the use of the main fucntions. Use browseVignettes(package="inctools") to access them.

This function summarizes performance of a recent infection test into a standard error of the incidence estimate, given the estimated test properties and hypothetical survey context or the sample size necessary for a given level of precision.

Up to two arguments can be specified as ranges, with the input parameter `step` specifying the number of increments between the endpoints of the two ranges supplied under the argument name. This yields output for each step. See the second and third example below for an illustration of this output.

Either the argument RSE_I or the argument n must be set to "out".

Examples

incprecision(I = 0.015, RSE_I = 0.25, PrevH = 0.2, CR = 1,
MDRI = 200, RSE_MDRI = 0.05, FRR = 0.01, RSE_FRR = 0.2,
BigT = 730, DE_H = 1.1, DE_R = 1, n = 'out')

#> $sample.size
#> [1] 3634
#> 
#> $Prev.HIV.and.recent
#> [1] 0.00833
#> 
#> $Prev.HIV.and.nonrecent
#> [1] 0.19167
#> 
#> $RSE.I.inf.sample
#> [1] 0.07606
#> 
#> $RSE.PrevH
#> [1] 0.0348
#> 
#> $RSE.PrevR
#> [1] 0.03558
#> 

incprecision(I = c(0.015,0.02), RSE_I = 0.25, PrevH = c(0.10,0.20),
CR = 1, MDRI = 200, RSE_MDRI = 0.05, FRR = 0.01, RSE_FRR = 0.2,
BigT = 700, DE_H = 1, DE_R = 1, n = 'out', step = 5)

#> $sample.size
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015          2660          2660         2660          2660        2660
#> I = 0.01625        2596          2596         2596          2596        2596
#> I = 0.0175         2547          2547         2547          2547        2547
#> I = 0.01875        2512          2512         2512          2512        2512
#> I = 0.02           2489          2489         2489          2489        2489
#> 
#> $Prev.HIV.and.recent
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015       0.00813       0.00813      0.00813       0.00813     0.00813
#> I = 0.01625     0.00876       0.00876      0.00876       0.00876     0.00876
#> I = 0.0175      0.00936       0.00936      0.00936       0.00936     0.00936
#> I = 0.01875     0.00992       0.00992      0.00992       0.00992     0.00992
#> I = 0.02        0.01045       0.01045      0.01045       0.01045     0.01045
#> 
#> $Prev.HIV.and.nonrecent
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015       0.09187       0.09187      0.09187       0.09187     0.09187
#> I = 0.01625     0.11624       0.11624      0.11624       0.11624     0.11624
#> I = 0.0175      0.14064       0.14064      0.14064       0.14064     0.14064
#> I = 0.01875     0.16508       0.16508      0.16508       0.16508     0.16508
#> I = 0.02        0.18955       0.18955      0.18955       0.18955     0.18955
#> 
#> $RSE.I.inf.sample
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015       0.05583       0.05583      0.05583       0.05583     0.05583
#> I = 0.01625     0.05798       0.05798      0.05798       0.05798     0.05798
#> I = 0.0175      0.06033       0.06033      0.06033       0.06033     0.06033
#> I = 0.01875     0.06285       0.06285      0.06285       0.06285     0.06285
#> I = 0.02        0.06549       0.06549      0.06549       0.06549     0.06549
#> 
#> $RSE.PrevH
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015       0.05817       0.05817      0.05817       0.05817     0.05817
#> I = 0.01625     0.05193       0.05193      0.05193       0.05193     0.05193
#> I = 0.0175      0.04717       0.04717      0.04717       0.04717     0.04717
#> I = 0.01875     0.04332       0.04332      0.04332       0.04332     0.04332
#> I = 0.02        0.04009       0.04009      0.04009       0.04009     0.04009
#> 
#> $RSE.PrevR
#>             PrevH = 0.1 PrevH = 0.125 PrevH = 0.15 PrevH = 0.175 PrevH = 0.2
#> I = 0.015       0.02061       0.02061      0.02061       0.02061     0.02061
#> I = 0.01625     0.02527       0.02527      0.02527       0.02527     0.02527
#> I = 0.0175      0.02975       0.02975      0.02975       0.02975     0.02975
#> I = 0.01875     0.03404       0.03404      0.03404       0.03404     0.03404
#> I = 0.02        0.03817       0.03817      0.03817       0.03817     0.03817
#> 

incprecision(I = 0.017, RSE_I = 'out', PrevH = c(0.10,0.20),
CR = 1, MDRI = 211, RSE_MDRI = 0.05, FRR = 0.009, RSE_FRR = 0.2,
BigT = 720, n = 5000, step = 5)

#> $RSE_I
#>           PrevH   RSE_I
#> 1   PrevH = 0.1 0.16868
#> 2 PrevH = 0.125 0.17352
#> 3  PrevH = 0.15 0.17885
#> 4 PrevH = 0.175 0.18470
#> 5   PrevH = 0.2 0.19112
#> 
#> $Prev.HIV.and.recent
#>           PrevH Prev.HIV.and.recent
#> 1   PrevH = 0.1             0.00947
#> 2 PrevH = 0.125             0.00945
#> 3  PrevH = 0.15             0.00944
#> 4 PrevH = 0.175             0.00943
#> 5   PrevH = 0.2             0.00942
#> 
#> $Prev.HIV.and.nonrecent
#>           PrevH Prev.HIV.and.nonrecent
#> 1   PrevH = 0.1                0.09053
#> 2 PrevH = 0.125                0.11555
#> 3  PrevH = 0.15                0.14056
#> 4 PrevH = 0.175                0.16557
#> 5   PrevH = 0.2                0.19058
#> 
#> $RSE.I.inf.sample
#>           PrevH RSE.I.inf.sample
#> 1   PrevH = 0.1          0.05363
#> 2 PrevH = 0.125          0.05557
#> 3  PrevH = 0.15          0.05824
#> 4 PrevH = 0.175          0.06166
#> 5   PrevH = 0.2          0.06585
#> 
#> $RSE.PrevH
#>           PrevH RSE.PrevH
#> 1   PrevH = 0.1   0.04243
#> 2 PrevH = 0.125   0.03742
#> 3  PrevH = 0.15   0.03367
#> 4 PrevH = 0.175   0.03071
#> 5   PrevH = 0.2   0.02828
#> 
#> $RSE.PrevR
#>           PrevH RSE.PrevR
#> 1   PrevH = 0.1   0.01383
#> 2 PrevH = 0.125   0.01748
#> 3  PrevH = 0.15   0.02113
#> 4 PrevH = 0.175   0.02479
#> 5   PrevH = 0.2   0.02845
#>