Incidence and incidence difference statistics from trinomial prevalences of HIV and recency

Incidence and incidence difference statistics from trinomial prevalences of HIV and recency

incprops(
  PrevH,
  RSE_PrevH,
  PrevR,
  RSE_PrevR,
  Boot = FALSE,
  BS_Count = 10000,
  alpha = 0.05,
  BMest = "same.test",
  MDRI,
  RSE_MDRI,
  FRR,
  RSE_FRR,
  BigT = 730,
  Covar_HR = 0,
  debug = FALSE
)

Arguments

PrevH
Prevalence of HIV (vector/integer).

RSE_PrevH
Relative Standard Error (RSE) of estimate for population prevalence of HIV (vector/integer).

PrevR
Proportion of persons found to be 'recent' by biomarker assay among total persons found positive for HIV (vector/integer).

RSE_PrevR
Relative Standard Error (RSE) of estimate for population proportion of those testing positive for HIV who have been infected recently (vector/integer).

Boot
True/False variable indicating whether variance of point estimates is to be calculated by Empirical Bootstrapping (TRUE) or Delta Method (FALSE), the default setting.

BS_Count
Specifies number of bootstrap samples for bootstrapped confidence intervals of incidence.

alpha
test rejection threshold.

BMest
Biomarker estimation by one the 3 options 'same.test'(=default), 'FRR.indep', 'MDRI.FRR.indep' (string).

MDRI
mean duration of recent infection [days] (vector/integer).

RSE_MDRI
Relative standard error of MDRI [days] (vector/integer).

FRR
False recent rate (vector/integer).

RSE_FRR
Relative standard error of FRR (vector/integer).

BigT
post-infection time cut-off true vs false recent [days] default 730 days (integer).

Covar_HR
Covariance of probability of being positive and being categorized recent from survey (vector/integer). Note that as the variances of PrevH and PrevR are often quite small, only a suitably commensurate covariance will enable the inversion of the bootstrap covariance matrix for random number generation to proceed without error.

debug
Enable debugging mode (browser)

Value

Returns incidence and incidence difference statistics, annualised risk and implied CIs on test characteristics.

Details

Implements assay-based incidence estimation through cross-sectional prevalence and recency of infection tests as described by Kassanjee, et al. 'A new general biomarker-based incidence estimator,' Epidemiology (2012). Function parameters must be specified to include assay test characteristics and survey results as proportions. Confidence intervals are computed via Delta method approximation, except when Boot=TRUE is specified, in which case confidence intervals are generated by empirical bootstrap resampling. Inputs must be in appropriate ranges for appropriate units. Extreme input values may make calculation impossible, and if entered will elicit error notices.

The package contains long form documentation in the form of vignettes that cover the use of the main fucntions. Use browseVignettes(package="inctools") to access them.

Examples

incprops(PrevH = 0.20, RSE_PrevH = 0.028, PrevR = 0.10, RSE_PrevR = 0.09,
         BS_Count = 10000, Boot = TRUE, MDRI = 200, RSE_MDRI = 0.05,
         FRR = 0.01,
         RSE_FRR = 0.2, BigT = 730)

#> $Incidence.Statistics
#> # A tibble: 1 x 6
#>   Incidence     CI_LB     CI_UB      RSE   Cov.PrevH.I Cor.PrevH.I
#>       <dbl>     <dbl>     <dbl>    <dbl>         <dbl>       <dbl>
#> 1 0.0426472 0.0330546 0.0531280 0.119692 0.00000831567    0.291931
#> 
#> $Annual.Risk.of.Infection
#> # A tibble: 1 x 3
#>         ARI ARI.CI_LB ARI.CI_UB
#>       <dbl>     <dbl>     <dbl>
#> 1 0.0417506 0.0325143 0.0517414
#> 
#> $MDRI.CI
#>      CI_LB    CI_UB
#> 1 180.4004 219.5996
#> 
#> $FRR.CI
#> # A tibble: 1 x 2
#>        CI_LB     CI_UB
#>        <dbl>     <dbl>
#> 1 0.00608007 0.0139199
#> 

incprops(PrevH = c(0.20,0.21,0.18), RSE_PrevH = c(0.028,0.03,0.022),
         PrevR = c(0.10,0.13,0.12), RSE_PrevR = c(0.094,0.095,0.05),
         BS_Count = 10000, Boot = FALSE, BMest = 'MDRI.FRR.indep', MDRI = c(200,180,180),
         RSE_MDRI = c(0.05,0.07,0.06), FRR = c(0.01,0.009,0.02), RSE_FRR = c(0.2,0.2,0.1),
         BigT = 730)

#> $Incidence.Statistics
#> # A tibble: 3 x 6
#>   survey Incidence     CI_LB     CI_UB       RSE RSE_InfSS
#>    <int>     <dbl>     <dbl>     <dbl>     <dbl>     <dbl>
#> 1      1 0.0426472 0.0323959 0.0528986 0.122642  0.0539212
#> 2      2 0.0677397 0.0503319 0.0851475 0.131115  0.0730176
#> 3      3 0.0484745 0.0396085 0.0573405 0.0933180 0.0662453
#> 
#> $Incidence.Difference.Statistics
#> # A tibble: 6 x 8
#>   compare        Diff        CI_LB        CI_UB      RSE RSE_InfSS   p_value
#>   <chr>         <dbl>        <dbl>        <dbl>    <dbl>     <dbl>     <dbl>
#> 1 1 vs 2  -0.0250925  -0.0452945   -0.00489049  0.410774  0.217381 0.0149152
#> 2 1 vs 3  -0.00582725 -0.0193807    0.00772616  1.18669   0.677794 0.399407 
#> 3 2 vs 1   0.0250925   0.00489049   0.0452945   0.410774  0.217381 0.0149152
#> 4 2 vs 3   0.0192652  -0.000270292  0.0388008   0.517372  0.306104 0.0532552
#> 5 3 vs 1   0.00582725 -0.00772616   0.0193807   1.18669   0.677794 0.399407 
#> 6 3 vs 2  -0.0192652  -0.0388008    0.000270292 0.517372  0.306104 0.0532552
#> # … with 1 more variable: p_value_InfSS <dbl>
#> 
#> $MDRI.CI
#>      CI_LB    CI_UB
#> 1 180.4004 219.5996
#> 2 155.3045 204.6955
#> 3 158.8324 201.1676
#> 
#> $FRR.CI
#> # A tibble: 3 x 2
#>        CI_LB     CI_UB
#>        <dbl>     <dbl>
#> 1 0.00608007 0.0139199
#> 2 0.00547206 0.0125279
#> 3 0.0160801  0.0239199
#>